requirements.txt
# ML Framework
torch>=2.0.0
torchvision>=0.15.0

# MLOps - Experiment Tracking
mlflow>=2.10.0

# API Serving
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Data & Utils
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Code Quality
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
